Development

+++ checker/src/main/kotlin/domain/TestCasesRepository.kt
@@ -1,9 +1,9 @@
 package com.sushkpavel.domain
 
-import com.sushkpavel.domain.model.TestCase
-import com.sushkpavel.domain.model.TestCaseDTO
+import com.sushkpavel.tasktester.entities.checker.TestCase
+import com.sushkpavel.infrastructure.model.TestCaseDTO
 
 interface TestCasesRepository {
-    suspend fun getTestCasesByTaskId(taskId: Int): List<TestCase>
+    suspend fun getTestCasesByTaskId(taskId: Long): List<TestCase>
     suspend fun postTestCases(testCaseDTO: TestCaseDTO): Int
 }
