+++ checker/src/main/kotlin/domain/service/CheckerService.kt
@@ -1,13 +1,13 @@
 package com.sushkpavel.domain.service
 
-import com.sushkpavel.domain.model.SubmissionRequest
-import com.sushkpavel.domain.model.TestCase
-import com.sushkpavel.domain.model.TestCaseDTO
+import com.sushkpavel.infrastructure.model.SubmissionRequest
+import com.sushkpavel.tasktester.entities.checker.TestCase
+import com.sushkpavel.infrastructure.model.TestCaseDTO
 import com.sushkpavel.domain.model.TestResult
 
 interface CheckerService {
     suspend fun checkTask(subRequest : SubmissionRequest) : TestResult
-    suspend fun getTestCases(taskId: Int): List<TestCase>
+    suspend fun getTestCases(taskId: Long): List<TestCase>
 
     suspend fun postTestCases(testCaseDTO: TestCaseDTO) : Int
 }
