+++ submission/src/main/java/com/sushkpavel/domain/repo/TestTaskRepository.kt
@@ -1,13 +1,13 @@
 package com.sushkpavel.domain.repo
 
-import com.sushkpavel.domain.dto.SubmissionRequest
-import com.sushkpavel.domain.dto.TestResultDTO
-import com.sushkpavel.domain.model.SolutionSubmission
-import com.sushkpavel.domain.model.TestResult
+import com.sushkpavel.infrastructure.dto.SubmissionRequest
+import com.sushkpavel.infrastructure.dto.TestResultDTO
+import com.sushkpavel.tasktester.entities.submission.SolutionSubmission
+import com.sushkpavel.tasktester.entities.submission.TestResult
 
 interface TestTaskRepository {
     suspend fun testTask(submissionRequest: SubmissionRequest) : TestResultDTO
     suspend fun saveResult(testResultDTO: TestResultDTO, submission: SolutionSubmission)
-    suspend fun getResultsForTask(userId : Int, taskId : Int) : List<TestResult?>
+    suspend fun getResultsForTask(userId : Int, taskId : Long) : List<TestResult?>
     suspend fun getResultForTask(userId : Int, submissionId : Int) : TestResult?
 }
